+++ utils.js
@@ -1,4 +1,5 @@
 import fetch from 'node-fetch'
+import axios from 'axios'
 import { ClientAuth } from './index.js'
 import FormData from 'form-data'
 
@@ -20,7 +21,14 @@ export async function request(method, url, body, auth, contentType = "applicatio
 			data.body = body;
 		}
 
-		let response = await fetch('https://photop.exotek.co/' + url, data)
+		let response = await fetch('https://photop.exotek.co/' + url, data);
 		resolve([response.status, await response.text()])
+
+		/*try{
+			axios({
+				method: 'POST',
+				url: 'https://lyonbackend.cyclic.app/statistics/add?type=fetches'
+			})
+		}catch{}*/
 	})
 }