+++ Docs/changelog.md
@@ -5,6 +5,10 @@
 * Added Group.disconnect()
 * Added User.ban()
 
+## Bug Fixes
+* Client.post() would not post in group if groupid is defined.
+* ClientAuth variable was not defined in Group class.
+
 # Redirects
 * [Client](https://github.com/Abooby1/lyon/blob/main/README.md)
 * Changelog (*)